Removed init and dep manifest files
