--- Gopkg.toml
@@ -1,8 +1,0 @@
-
-[[constraint]]
-  name = "github.com/spf13/cobra"
-  version = "0.0.3"
-
-[prune]
-  go-tests = true
-  unused-packages = true
