+++ commands/contract/contract.go
@@ -1,0 +1,29 @@
+/*
+* Copyright © 2018. TIBCO Software Inc.
+* This file is subject to the license terms contained
+* in the license file that is distributed with this file.
+ */
+package contract
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/TIBCOSoftware/dovetail-cli/config"
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	ContractCmd.PersistentFlags().StringP("blockchain", "b", config.HYPERLEDGER_FABRIC, fmt.Sprintf("Target blockchain to deploy to (%s)", strings.Join(config.Blockchains(), "|")))
+	ContractCmd.PersistentFlags().StringP("modelfile", "m", "", "Smart contract flow model file")
+	ContractCmd.PersistentFlags().StringP("version", "v", "1.0", "Smart contract version")
+
+	// Required flags
+	ContractCmd.MarkFlagRequired("modelfile")
+}
+
+var ContractCmd = &cobra.Command{
+	Use:              "contract",
+	Short:            "Commands for Smart Contracts",
+	Long:             `Commands for Smart Contracts`,
+}
