+++ commands/contract/generate.go
@@ -11,7 +11,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/TIBCOSoftware/dovetail-cli/cmd"
 	"github.com/TIBCOSoftware/dovetail-cli/config"
 	corda "github.com/TIBCOSoftware/dovetail-cli/corda/contract"
 	fabric "github.com/TIBCOSoftware/dovetail-cli/hyperledger-fabric/contract"
@@ -30,7 +29,7 @@ var modelfile string
 var enableTxnSecurity bool
 
 func init() {
-	contractCmd.AddCommand(generateCmd)
+	ContractCmd.AddCommand(generateCmd)
 	generateCmd.PersistentFlags().StringP("target", "t", ".", "Destination path for generated artifacts")
 	generateCmd.Flags().StringP("state", "", "", "Corda only, optional, specify asset name to generate contract state, default to all assets in the specified namespace")
 	generateCmd.Flags().StringP("commands", "", "", "Corda only, optional, comma delimited list of transactions(commands) allowed for the selected state txn1,txn2,..., default to all transactions")
@@ -44,20 +43,19 @@ var generateCmd = &cobra.Command{
 	Use:              "generate",
 	Short:            "Commands for generating smart contract artifacts",
 	Long:             `Commands for generating smart contract artifacts`,
-	PersistentPreRun: cmd.ConfigPreRun,
 	Run: func(cmd *cobra.Command, args []string) {
-		blockchain, err := contractCmd.PersistentFlags().GetString("blockchain")
+		blockchain, err := ContractCmd.PersistentFlags().GetString("blockchain")
 		if err != nil {
 			fmt.Println(err)
 			os.Exit(1)
 		}
 
-		smversion, err = contractCmd.PersistentFlags().GetString("version")
+		smversion, err = ContractCmd.PersistentFlags().GetString("version")
 		if err != nil {
 			fmt.Println(err)
 			os.Exit(1)
 		}
-		modelfile, err = contractCmd.PersistentFlags().GetString("modelfile")
+		modelfile, err = ContractCmd.PersistentFlags().GetString("modelfile")
 		if err != nil {
 			fmt.Println(err)
 			os.Exit(1)
